Fix #4740, #4749, #4751 - NPE crash, power hook return type, assembly table refresh
- Fix #4740: Add null check in GuiUtil.getUnFormattedTooltip() to prevent NPE
  when ItemStack.getItem() returns null for certain mod items

- Fix #4749: Change IPipeTransportPowerHook return types from int to long,
  fix requestPower -1 fallback logic, and add missing receivePower hook call

- Fix #4751: Override onSlotChange() in TileAssemblyTable to trigger
  immediate recipe update when items are swapped in the GUI

+++ common/buildcraft/transport/pipe/flow/IPipeTransportPowerHook.java
@@ -13,8 +13,10 @@ public interface IPipeTransportPowerHook {
     /** Override default behavior on receiving energy into the pipe.
      *
      * @return The amount of power used, or -1 for default behavior. */
-    int receivePower(EnumFacing from, long val);
+    long receivePower(EnumFacing from, long val);
 
-    /** Override default requested power. */
-    int requestPower(EnumFacing from, long amount);
+    /** Override default requested power.
+     *
+     * @return The amount of power requested, or -1 for default behavior. */
+    long requestPower(EnumFacing from, long amount);
 }
